+++ frontend/pages/home/helpers/icrc/getIcrcCanister.ts
@@ -19,3 +19,5 @@ export function getCanister(assetAddress: string | Principal): IcrcLedgerCaniste
         canisterId,
     });
 }
+
+
